requirements.txt
cv2
mediapipe==0.10.21
pyserial